+++ tests/testthat/test_getMappableRegions.R
@@ -27,7 +27,7 @@ test_that(".writeWindowsToTempFile works properly", {
 })
 
 test_that(".writeWindowsToTempFileCPP works properly", {
-    
+
     chrs1 <- as.character(XVector::subseq(chrs[[1]], start = 1, width = 100))
     expect_error(.writeWindowsToTempFileCPP(chr = chrs1, w = 10, fname = file.path(tempdir(), "error", "error.fa")))
     tf1 <- tempfile()
@@ -37,7 +37,7 @@ test_that(".writeWindowsToTempFileCPP works properly", {
     expect_identical(lns1[c(1,2,181,182)],
                      c(">1", substr(chrs1, 1, 10),
                        ">91", substr(chrs1, 91, 100)))
-    
+
     expect_type(res2 <- .writeWindowsToTempFileCPP(chr = chrs1, w = 100, tempfile()), "character")
     expect_identical(length(lns2 <- readLines(res2)), 2L)
     expect_identical(lns2, c(">1", chrs1))
@@ -51,10 +51,10 @@ test_that(".writeWindowsToTempFileCPP works properly", {
 
 test_that(".alignWindowsToGenome works properly", {
     skip_if_not_installed("Rbowtie")
-    
+
     chrs1 <- as.character(XVector::subseq(chrs[[1]], start = 2600, width = 1000))
     tf1 <- .writeWindowsToTempFile(chr = chrs1, w = 50)
-    
+
     tfs <- c(tempfile(), tempfile(), tempfile())
     expect_type(res1 <- .alignWindowsToGenome(tf1, indexname, m = 1, p = 2),
               "character")
@@ -83,40 +83,27 @@ test_that("getMappableRegions() works properly", {
 
     skip_if_not_installed("QuasR")
     skip_if_not_installed("BSgenome")
-
-    library("BSgenome")
+    skip_if_not_installed("withr")
 
     # copy sample data
     file.copy(system.file("extdata", package = "QuasR"), tempdir(), recursive = TRUE)
-    
-    # temporary R library
-    rlibdir <- tempfile(pattern = "Rlib")
-    dir.create(rlibdir)
-    .libPaths(rlibdir)
-    # ... add rlibdir to R_LIBS for "R CMD INSTALL" and cluster nodes to find it
-    oldRlibs <- Sys.getenv("R_LIBS")
-    Sys.setenv(R_LIBS = paste(tools::file_path_as_absolute(rlibdir), oldRlibs,
-                              sep = .Platform$path.sep))
-    
-    # create cluster object
+
+    # -------------- temporary R library
+    withr::local_temp_libpaths()
+    withr::local_envvar(c(R_LIBS = paste(.libPaths(), collapse = .Platform$path.sep)))
+
+    # create BSgenome and Rbowtie index
     clObj <- parallel::makeCluster(2L)
-    
-    # load QuasR on cluster nodes
-    parallel::clusterEvalQ(cl = clObj, expr = library(QuasR))
-    parallel::clusterEvalQ(cl = clObj, expr = library(BSgenome))
-    
-    # install BSgenome.HSapiens.QuasR.hg19sub into temporary library
     bsgPkg <- file.path(tempdir(), "extdata", "BSgenome.HSapiens.QuasR.hg19sub_0.1.0.tar.gz")
     suppressMessages({
-        utils::install.packages(pkgs = bsgPkg, lib = rlibdir, repos = NULL,
+        utils::install.packages(pkgs = bsgPkg, repos = NULL,
                                 type = "source", INSTALL_opts = "--no-test-load")
     })
-    
-    # build Rbowtie index
     samplefile <- file.path(tempdir(), "extdata", "samples_chip_single.txt")
     genomePkg <- "BSgenome.HSapiens.QuasR.hg19sub"
-    proj <- QuasR::qAlign(samplefile, genomePkg, clObj = clObj, lib.loc = rlibdir)
-        
+    proj <- QuasR::qAlign(samplefile, genomePkg, clObj = clObj, lib.loc = .libPaths()[1])
+    parallel::stopCluster(cl = clObj)
+
     # test getMappableRegions
     expect_s4_class(gr2 <- getMappableRegions(genomePkg, indexname, 50), "GRanges")
     expect_identical(gr1, gr2)
@@ -124,6 +111,9 @@ test_that("getMappableRegions() works properly", {
     expect_identical(gr1, gr3)
     expect_s4_class(gr4 <- getMappableRegions(get(genomePkg), indexname, 50), "GRanges")
     expect_identical(gr1, gr4)
+
+    # -------------- restore original library paths
+    withr::deferred_run()
 
     unlink(readfile)
 })